+++ src/parse/cpu.rs
@@ -1,6 +1,4 @@
-use super::{elementext::ElementExt, Element, Parse};
-
-use crate::error::*;
+use super::{elementext::ElementExt, Element, Parse, Result, SVDError};
 use crate::svd::{Cpu, Endian};
 
 impl Parse for Cpu {
