+++ src/parse/dimelement.rs
@@ -1,9 +1,4 @@
-use super::{
-    elementext::ElementExt,
-    types::{parse_optional, DimIndex},
-    Element, Parse,
-};
-use crate::error::*;
+use super::{elementext::ElementExt, optional, types::DimIndex, Element, Parse, Result};
 use crate::svd::DimElement;
 
 impl Parse for DimElement {
@@ -14,7 +9,7 @@ impl Parse for DimElement {
         Ok(DimElement::builder()
             .dim(tree.get_child_u32("dim")?)
             .dim_increment(tree.get_child_u32("dimIncrement")?)
-            .dim_index(parse_optional::<DimIndex>("dimIndex", tree)?)
+            .dim_index(optional::<DimIndex>("dimIndex", tree)?)
             .build()?)
     }
 }
