+++ src/parse/interrupt.rs
@@ -1,6 +1,4 @@
-use super::{elementext::ElementExt, Element, Parse};
-
-use crate::error::*;
+use super::{elementext::ElementExt, Context, Element, Parse, Result, SVDError};
 use crate::svd::Interrupt;
 
 impl Interrupt {
