+++ src/parse/types.rs
@@ -3,11 +3,8 @@
 
 use xmltree::Element;
 
-pub use crate::parse::optional as parse_optional;
-pub use crate::parse::Parse;
-
 use super::elementext::ElementExt;
-use crate::error::*;
+use super::{Context, Parse, Result, SVDError};
 
 macro_rules! unwrap {
     ($e:expr) => {
