+++ src/parse/usage.rs
@@ -1,6 +1,4 @@
-use super::{elementext::ElementExt, Element, Parse};
-
-use crate::error::*;
+use super::{elementext::ElementExt, Element, Parse, Result, SVDError};
 use crate::svd::Usage;
 
 impl Parse for Usage {
